+++ src/Oro/Bundle/ChannelBundle/Tests/Unit/Provider/EntityExclusionProviderTest.php
@@ -2,113 +2,100 @@
 
 namespace Oro\Bundle\ChannelBundle\Tests\Unit\Provider;
 
+use Doctrine\ORM\Mapping\ClassMetadata;
 use Oro\Bundle\ChannelBundle\Provider\EntityExclusionProvider;
 use Oro\Bundle\ChannelBundle\Provider\SettingsProvider;
 use Oro\Bundle\ChannelBundle\Provider\StateProvider;
 
 class EntityExclusionProviderTest extends \PHPUnit\Framework\TestCase
 {
-    const TEST_MAIN_ENTITY_NAME               = 'TestBundle\Entity\TestMain';
-    const TEST_ENTITY_NAME                    = 'TestBundle\Entity\Test';
-    const TEST_DEPENDENCY_ENTITY_NAME         = 'TestBundle\Entity\Test1';
-    const TEST_ANOTHER_DEPENDENCY_ENTITY_NAME = 'TestBundle\Entity\Test12';
-    const TEST_ASSOC_NAME                     = 'relation';
-    const TEST_FIELD_NAME                     = 'field';
-
-    /** @var EntityExclusionProvider */
-    protected $exclusionProvider;
+    private const TEST_MAIN_ENTITY_NAME               = 'TestBundle\Entity\TestMain';
+    private const TEST_ENTITY_NAME                    = 'TestBundle\Entity\Test';
+    private const TEST_DEPENDENCY_ENTITY_NAME         = 'TestBundle\Entity\Test1';
+    private const TEST_ANOTHER_DEPENDENCY_ENTITY_NAME = 'TestBundle\Entity\Test12';
+    private const TEST_ASSOC_NAME                     = 'relation';
+    private const TEST_FIELD_NAME                     = 'field';
 
     /** @var SettingsProvider|\PHPUnit\Framework\MockObject\MockObject */
-    protected $settingsProvider;
+    private $settingsProvider;
 
     /** @var StateProvider|\PHPUnit\Framework\MockObject\MockObject */
-    protected $stateProvider;
+    private $stateProvider;
 
-    public function setUp()
+    /** @var EntityExclusionProvider */
+    private $exclusionProvider;
+
+    protected function setUp()
     {
-        $this->settingsProvider = $this->getMockBuilder('Oro\Bundle\ChannelBundle\Provider\SettingsProvider')
-            ->disableOriginalConstructor()
-            ->setMethods(['isChannelEntity', 'getDependentEntityData'])
-            ->getMock();
-        $this->stateProvider    = $this->getMockBuilder('Oro\Bundle\ChannelBundle\Provider\StateProvider')
-            ->disableOriginalConstructor()->getMock();
+        $this->settingsProvider = $this->createMock(SettingsProvider::class);
+        $this->stateProvider = $this->createMock(StateProvider::class);
 
         $this->exclusionProvider = new EntityExclusionProvider($this->settingsProvider, $this->stateProvider);
     }
 
-    public function tearDown()
-    {
-        unset($this->exclusionProvider, $this->stateProvider, $this->settingsProvider);
-    }
-
     /**
      * @dataProvider exclusionProvider
-     *
-     * @param bool  $expected
-     * @param bool  $isChannelEntity
-     * @param array $dependentDataMap
-     * @param array $enabledEntityMap
      */
     public function testIsEntityExcluded(
-        $expected,
-        $isChannelEntity,
-        $dependentDataMap = [],
-        $enabledEntityMap = []
+        bool $expected,
+        bool $isChannelEntity,
+        array $dependentDataMap = [],
+        array $enabledEntityMap = []
     ) {
-        $this->settingsProvider->expects($this->any())
-            ->method('isChannelEntity')->with($this->equalTo(self::TEST_ENTITY_NAME))
-            ->will($this->returnValue($isChannelEntity));
-        $this->settingsProvider->expects($this->any())
-            ->method('getDependentEntityData')
-            ->will($this->returnValueMap($dependentDataMap));
-        $this->stateProvider->expects($this->any())
-            ->method('isEntityEnabled')
-            ->will($this->returnValueMap($enabledEntityMap));
+        $isDependentOnChannelEntity = count($dependentDataMap) > 1
+            || self::TEST_ENTITY_NAME !== $dependentDataMap[0][0]
+            || [] !== $dependentDataMap[0][1];
 
-        $this->assertSame($expected, $this->exclusionProvider->isIgnoredEntity(self::TEST_ENTITY_NAME));
+        $this->settingsProvider->expects(self::once())
+            ->method('isChannelEntity')
+            ->with(self::TEST_ENTITY_NAME)
+            ->willReturn($isChannelEntity);
+        $this->settingsProvider->expects(self::any())
+            ->method('isDependentOnChannelEntity')
+            ->with(self::TEST_ENTITY_NAME)
+            ->willReturn($isDependentOnChannelEntity);
+        $this->settingsProvider->expects(self::any())
+            ->method('getDependentEntities')
+            ->willReturnMap($dependentDataMap);
+        $this->stateProvider->expects(self::any())
+            ->method('isEntityEnabled')
+            ->willReturnMap($enabledEntityMap);
+
+        self::assertSame($expected, $this->exclusionProvider->isIgnoredEntity(self::TEST_ENTITY_NAME));
     }
 
     /**
      * @dataProvider exclusionProvider
-     *
-     * @param bool  $expected
-     * @param bool  $isChannelEntity
-     * @param array $dependentDataMap
-     * @param array $enabledEntityMap
      */
     public function testIsRelationExcluded(
-        $expected,
-        $isChannelEntity,
-        $dependentDataMap = [],
-        $enabledEntityMap = []
+        bool $expected,
+        bool $isChannelEntity,
+        array $dependentDataMap = [],
+        array $enabledEntityMap = []
     ) {
-        $this->settingsProvider->expects($this->any())
+        $this->settingsProvider->expects(self::any())
             ->method('isChannelEntity')
-            ->will(
-                $this->returnValueMap(
-                    [
-                        [self::TEST_MAIN_ENTITY_NAME, true],
-                        [self::TEST_ENTITY_NAME, $isChannelEntity]
-                    ]
-                )
-            );
-        $this->settingsProvider->expects($this->any())
-            ->method('getDependentEntityData')
-            ->will($this->returnValueMap($dependentDataMap));
-        $this->stateProvider->expects($this->any())
+            ->willReturnMap([
+                [self::TEST_MAIN_ENTITY_NAME, true],
+                [self::TEST_ENTITY_NAME, $isChannelEntity]
+            ]);
+        $this->settingsProvider->expects(self::any())
+            ->method('getDependentEntities')
+            ->willReturnMap($dependentDataMap);
+        $this->stateProvider->expects(self::any())
             ->method('isEntityEnabled')
-            ->will($this->returnValueMap($enabledEntityMap));
+            ->willReturnMap($enabledEntityMap);
 
-        $classMetadataMock = $this->getMockBuilder('Doctrine\ORM\Mapping\ClassMetadata')
-            ->disableOriginalConstructor()->getMock();
-        $classMetadataMock->expects($this->once())
+        $classMetadataMock = $this->createMock(ClassMetadata::class);
+        $classMetadataMock->expects(self::any())
             ->method('getName')
-            ->will($this->returnValue(self::TEST_MAIN_ENTITY_NAME));
-        $classMetadataMock->expects($this->once())
-            ->method('getAssociationTargetClass')->with($this->equalTo(self::TEST_ASSOC_NAME))
-            ->will($this->returnValue(self::TEST_ENTITY_NAME));
+            ->willReturn(self::TEST_MAIN_ENTITY_NAME);
+        $classMetadataMock->expects(self::any())
+            ->method('getAssociationTargetClass')
+            ->with(self::TEST_ASSOC_NAME)
+            ->willReturn(self::TEST_ENTITY_NAME);
 
-        $this->assertSame(
+        self::assertSame(
             $expected,
             $this->exclusionProvider->isIgnoredRelation($classMetadataMock, self::TEST_ASSOC_NAME)
         );
@@ -116,10 +103,9 @@ class EntityExclusionProviderTest extends \PHPUnit\Framework\TestCase
 
     public function testIsFieldExcluded()
     {
-        $classMetadataMock = $this->getMockBuilder('Doctrine\ORM\Mapping\ClassMetadata')
-            ->disableOriginalConstructor()->getMock();
+        $classMetadataMock = $this->createMock(ClassMetadata::class);
 
-        $this->assertFalse(
+        self::assertFalse(
             $this->exclusionProvider->isIgnoredField($classMetadataMock, self::TEST_FIELD_NAME),
             'should not exclude any fields'
         );
@@ -134,19 +120,19 @@ class EntityExclusionProviderTest extends \PHPUnit\Framework\TestCase
             'not related to channel entity given, should not be skipped' => [
                 false,
                 false,
-                [[self::TEST_ENTITY_NAME, false]],
+                [[self::TEST_ENTITY_NAME, []]],
                 [[self::TEST_MAIN_ENTITY_NAME, true]]
             ],
             'channel entity given, but not enabled'                      => [
                 true,
                 true,
-                [[self::TEST_ENTITY_NAME, false]],
+                [[self::TEST_ENTITY_NAME, []]],
                 [[self::TEST_ENTITY_NAME, false], [self::TEST_MAIN_ENTITY_NAME, true]]
             ],
             'channel entity given, enabled entity should not be skipped' => [
                 false,
                 true,
-                [[self::TEST_ENTITY_NAME, false]],
+                [[self::TEST_ENTITY_NAME, []]],
                 [[self::TEST_ENTITY_NAME, true], [self::TEST_MAIN_ENTITY_NAME, true]]
             ],
             'dependent entity given, all dependencies disabled'          => [